Ordering all the mess, 1) separate the routes from app.py into routes.py. 2) create directory to the forms, and separate the forms into 2 files. 3) create another pages to get close for full end-to-end attack.
