--- Server/logicFuncs.py
@@ -1,7 +1,0 @@
-# Place where i will develop the functions that will be used in the server
-# funcs can be placed in relevant files ( HTML etc..)
-
-# TODO: Implement file upload and save for voice samples for model logic
-# TODO: func that handles and creates the prompts
-# TODO: func that handles conversation recording ( between two people )
-# TODO: func to handle model training/tuning/voice profile creater using uploaded voice samples
