--- Server/templates/403.html
@@ -1,9 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Flasky - Forbidden{% endblock %}
-
-{% block page_content %}
-<div class="page-header">
-    <h1>Forbidden</h1>
-</div>
-{% endblock %}
