--- Server/templates/500.html
@@ -1,9 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}Deceptify - Internal Server Error{% endblock %}
-
-{% block page_content %}
-<div class="page-header">
-    <h1>Internal Server Error</h1>
-</div>
-{% endblock %}
