+++ Server/templates/attack_pages/new_profile.html
@@ -1,0 +1,14 @@
+{% import "_macros.html" as macros %}
+{% extends "base.html" %}
+{% import "bootstrap/wtf.html" as wtf %}
+
+
+{% block title %}Deceptify- New Profile{%endblock%}
+
+{% block page_content %}
+<h2>
+    Create a New Profile.
+</h2>
+{{ wtf.quick_form(form) }}
+
+{%endblock%}
