+++ Server/templates/attack_pages/new_voice_attack.html
@@ -1,0 +1,13 @@
+{% import "_macros.html" as macros %}
+{% extends "base.html" %}
+{% import "bootstrap/wtf.html" as wtf %}
+
+
+{% block page_content %}
+<div class="page-header">
+    <h1>Generate New Voice Attack</h1>
+</div>
+
+{{ wtf.quick_form(form) }}
+
+{%endblock%}
