+++ Server/templates/attack_pages/newattack.html
@@ -1,0 +1,9 @@
+{% import "_macros.html" as macros %}
+{% extends "base.html" %}
+
+{% block page_content %}
+<div class="page-header">
+    <h1>New Attack</h1>
+</div>
+
+{%endblock%}
