+++ Server/templates/base.html
@@ -18,7 +18,6 @@
         <div class="navbar-collapse collapse">
             <ul class="nav navbar-nav">
                 <li><a href="{{url_for('index')}}">Home</a></li>
-                <li><a href="{{url_for('newchat')}}">New Chat</a></li>
                 <li><a href="{{url_for('dashboard')}}">Dash Board</a></li>
                 <li><a href="{{url_for('contact')}}">Contact Us</a></li>
