+++ Server/templates/contact.html
@@ -2,6 +2,8 @@
 {% extends "base.html" %}
 {% import "bootstrap/wtf.html" as wtf %}
 
+{% block title %}Deceptify- Contact Us{%endblock%}
+
 
 {% block page_content %}
 <div class="page-header">
