+++ Server/templates/data_collection_pages/collect_video.html
@@ -1,0 +1,13 @@
+{% import "_macros.html" as macros %}
+{% extends "base.html" %}
+{% import "bootstrap/wtf.html" as wtf %}
+
+{% block title %}Deceptify- Upload Video{%endblock%}
+
+
+{% block page_content %}
+<div class="page-header">
+    <h1>Upload the Video, Please.</h1>
+</div>
+{{ wtf.quick_form(form) }}
+{%endblock%}
