+++ Server/templates/data_collection_pages/upload_voice_file.html
@@ -1,0 +1,11 @@
+{% import "_macros.html" as macros %}
+{% extends "base.html" %}
+{% import "bootstrap/wtf.html" as wtf %}
+
+
+{% block page_content %}
+<div class="page-header">
+    <h1>Upload Existing Voice Recording File</h1>
+</div>
+{{ wtf.quick_form(form) }}
+{%endblock%}
