+++ Server/templates/errors/404.html
@@ -1,0 +1,9 @@
+{% extends "base.html" %}
+
+{% block title %}Deceptify - Page Not Found{% endblock %}
+
+{% block page_content %}
+<div class="page-header">
+    <h1>Not Found</h1>
+</div>
+{% endblock %}
