+++ Server/templates/index.html
@@ -8,12 +8,12 @@
     <h1>Welcome, User!</h1>
 </div>
 <div style="display: flex; justify-content: center; align-items: center; height: 100vh;">
-    <a href="{{url_for('new_voice_attack')}}">
-        <button type="button">Generate Voice Attack</button>
-    </a>
+<!--    <a href="{{url_for('new_voice_attack')}}">-->
+<!--        <button type="button">Generate Voice Attack</button>-->
+<!--    </a>-->
 
     <a href="{{url_for('information_gathering')}}">
-        <button type="button">Gather Information</button>
+        <button type="button">Start Attack Generation Process</button>
     </a>
 
     <a href="{{url_for('new_profile')}}">
