--- Server/templates/information_gathering.html
@@ -1,16 +1,0 @@
-{% import "_macros.html" as macros %}
-{% extends "base.html" %}
-{% import "bootstrap/wtf.html" as wtf %}
-
-
-{% block title %}Deceptify- Information Gathering{%endblock%}
-
-{% block page_content %}
-<p>At the first stage, we need you to upload data on the victim; after doing that, we can provide you the next steps in
-    the
-    attack generation.
-</p>
-
-{{ wtf.quick_form(form) }}
-
-{%endblock%}
