--- Server/templates/newchat.html
@@ -1,9 +1,0 @@
-{% import "_macros.html" as macros %}
-{% extends "base.html" %}
-
-{% block page_content %}
-<div class="page-header">
-    <h1>New Chat</h1>
-</div>
-
-{%endblock%}
